requirements.txt
Flask
python-dotenv
requests
pyTelegramBotAPI
mysqlclient
Flask-MySQLdb